# RESTAPI/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Trabalho PAW",
    "description": "REST API para a milestone 1"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
...
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/login": {
      "post": {
        "summary": "Fazer login na API",
        "tags": [
          "Sessão"
        ],
        "parameters": [
          {
            "name": "Dados de Login",
            "in": "body",
            "description": "Informações do utilizador",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Log in com sucesso"
          }
        }
      }
    },
...
    "/changePassword": {
      "put": {
        "summary": "Mudar password",
        "tags": [
          "Sessão"
        ],
        "parameters": [
          {
            "name": "Password",
            "in": "body",
            "description": "Password antiga e password nova",
            "schema": {
              "$ref": "#/definitions/password"
            }
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "login": {
      "required": [],
      "properties": {
        "nmrCC": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "password": {
      "required": [],
      "properties": {
        "currentPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      }
    }
  }
}